// src/types/brands.ts
export interface BrandItem {
  id: string
  name: string
  description: string
  image: string
  buttonText: string
  link: string
}

export interface YarnProduct {
  id: string
  name: string
  description: string
  icon: string
  image: string
}

export interface JourneyMilestone {
  year: string
  title: string
  description: string
}

export interface AchievementItem {
  title: string
  description: string
  icon: string
}